requirements.txt
streamlit
pandas
psycopg2-binary
plotly
streamlit-plotly-events
openpyxl
python-dateutil

psycopg2-binary
python-dotenv
